공지사항 페이징 관련 매퍼 사용하는 걸로 갱신

+++ src/main/java/app/finup/layer/domain/notice/mapper/NoticeMapper.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface NoticeMapper {
     List<NoticeDto.Row> search(NoticeDto.Search rq);
-    Long countForSearch(NoticeDto.Search rq);
+    Integer countForSearch(NoticeDto.Search rq);
 }
